+++ qoqo-qryd/tests/integration/tweezer_devices.rs
@@ -44,7 +44,7 @@ fn test_new() {
                 .unwrap()
                 .extract::<String>()
                 .unwrap(),
-            "qryd_tweezer_device"
+            "testdevice"
         );
         assert_eq!(
             res_mut
@@ -52,7 +52,7 @@ fn test_new() {
                 .unwrap()
                 .extract::<String>()
                 .unwrap(),
-            "qryd_tweezer_device"
+            "testdevice"
         );
         assert_eq!(
             res.call_method0("seed")
@@ -306,6 +306,18 @@ fn test_allowed_tweezer_shifts_from_rows() {
         assert!(device_mut
             .call_method1("set_allowed_tweezer_shifts_from_rows", (vec![vec![0, 0]],))
             .is_err());
+    })
+}
+
+/// Test allow_reset for TweezerMutableDeviceWrapper
+#[test]
+fn test_allow_reset() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let device_type_mut = py.get_type::<TweezerMutableDeviceWrapper>();
+        let device_mut = device_type_mut.call0().unwrap();
+
+        assert!(device_mut.call_method1("set_allow_reset", (true,)).is_ok());
     })
 }